+++ index.d.ts
@@ -8,23 +8,19 @@ export interface Session {
   [key: string]: any
 }
 
-export declare const APILoginRoute: (req: NextRequest) => Promise<Response>
+export declare const APICreateLoginRoute: (
+  findUser: FindUser,
+  validatePassword: ValidatePassword
+) => Promise<Response>
 export declare const APILogoutRoute: () => Promise<Response>
 export declare const APIRefreshSessionRoute: () => Promise<Response>
 export declare const getSession: () => Promise<Session>
-export declare const setLocalStrategy: (
-  findUser: FindUser,
-  validatePassword: ValidatePassword
-) => void
 
 declare const NextjsAppPassport: {
-  APILoginRoute: typeof APILoginRoute
+  APICreateLoginRoute: typeof APICreateLoginRoute
   APILogoutRoute: typeof APILogoutRoute
   APIRefreshSessionRoute: typeof APIRefreshSessionRoute
   getSession: typeof getSession
-  setLocalStrategy: typeof setLocalStrategy
-  //@deprecated use setLocalStrategy instead
-  setLocaLStrategy: typeof setLocalStrategy
 }
 
 export default NextjsAppPassport