+++ src/index.ts
@@ -1,31 +1,23 @@
-import { loginRoute } from '@/app/api/login'
+import { createLoginRoute } from '@/app/api/login'
 import { logoutRoute } from '@/app/api/logout'
 import { refreshSessionRoute } from '@/app/api/refreshSession'
 
 import { getSession as getSession0 } from '@/lib/session'
-import { setLocalStrategy as setLocalStrategy0 } from '@/lib/strategy'
 
 // API routes
-export const APILoginRoute = loginRoute
+export const APICreateLoginRoute = createLoginRoute
 export const APILogoutRoute = logoutRoute
 export const APIRefreshSessionRoute = refreshSessionRoute
 
 // Session
 export const getSession = getSession0
 
-// Strategy
-export const setLocalStrategy = setLocalStrategy0
-//@deprecated use setLocalStrategy instead
-export const setLocaLStrategy = setLocalStrategy0
-
 // Default
 const NextjsAppPassport = {
-  APILoginRoute,
+  APICreateLoginRoute,
   APILogoutRoute,
   APIRefreshSessionRoute,
-  getSession,
-  setLocalStrategy,
-  setLocaLStrategy
+  getSession
 }
 
 export default NextjsAppPassport