+++ tests/index.test.ts
@@ -2,10 +2,14 @@ import { NextRequest } from 'next/server'
 
 import NextjsAppPassport from '@/index'
 
-const mockLogin = jest.fn()
-jest.mock('@/app/api/login', () => ({
-  loginRoute: async () => mockLogin()
-}))
+const mockCreateLogin = jest.fn()
+jest.mock(
+  '@/app/api/login',
+  () =>
+    (...args: any) =>
+    async (req: NextRequest) =>
+      mockCreateLogin(...args, req)
+)
 const mockLogout = jest.fn()
 jest.mock('@/app/api/logout', () => ({
   logoutRoute: async () => mockLogout()
@@ -18,10 +22,6 @@ const mockGetSession = jest.fn()
 jest.mock('@/lib/session', () => ({
   getSession: async () => mockGetSession()
 }))
-const mockStrategy = jest.fn()
-jest.mock('@/lib/strategy', () => ({
-  setLocalStrategy: () => mockStrategy()
-}))
 
 describe('@/index', () => {
   const req = {} as NextRequest
@@ -29,16 +29,13 @@ describe('@/index', () => {
   const validatePassword = jest.fn()
 
   test('default', async () => {
-    await NextjsAppPassport.APILoginRoute(req)
-    expect(mockLogin).toHaveBeenCalledTimes(1)
+    await NextjsAppPassport.APICreateLoginRoute(findUser, validatePassword)(req)
+    expect(mockCreateLogin).toHaveBeenCalledTimes(1)
 
     await NextjsAppPassport.APILogoutRoute(req)
     expect(mockLogout).toHaveBeenCalledTimes(1)
 
     await NextjsAppPassport.APIRefreshSessionRoute(req)
     expect(mockRefreshSession).toHaveBeenCalledTimes(1)
-
-    NextjsAppPassport.setLocalStrategy(findUser, validatePassword)
-    expect(mockStrategy).toHaveBeenCalledTimes(1)
   })
 })