+++ tests/lib/strategy/index.test.ts
@@ -1,4 +1,3 @@
-import { setLocalStrategy } from '@/lib/strategy'
 import { NextRequest } from 'next/server'
 import passport from 'passport'
 
@@ -46,7 +45,6 @@ describe('@/lib/strategy', () => {
   })
 
   test('success', async () => {
-    setLocalStrategy(findUser, validatePassword)
     const user = await authenticate(req)
     expect(mockFindUser).toHaveBeenCalledTimes(1)
     expect(mockFindUser).toHaveBeenCalledWith({